fix(project): include all enabled agents in preset default export targets
Root cause of #400: getDefaultExportAgents returned
prioritized.length > 0 ? prioritized : fallback, so whenever any
priority-list agent (claude_code/codex/cursor/gemini_cli/github_copilot)
was enabled, every non-priority agent (e.g. pi) was silently dropped
from the default export targets, and preset sync in the project view
never reached it. The trailing .slice(0, 3) also silently capped the
target list at three agents.

Fix: keep priority agents first, then append every remaining
enabled+installed agent in detected order, and drop the 3-agent cap.
The pure helper is extracted to src/lib/exportAgents.ts so it can be
exercised directly.
